Add relative path handling to `application.navigateToUrl` (#78565)
* split application utilities and associated tests to distinct files

* do not match app if path does not start with the basePath

* add relative paths support to `navigateToUrl`

* add null-check error

* update generated doc

* nits on doc
