+++ docs/development/core/public/kibana-plugin-core-public.applicationstart.geturlforapp.md
@@ -4,9 +4,11 @@
 
 ## ApplicationStart.getUrlForApp() method
 
-Returns an URL to a given app, including the global base path. By default, the URL is relative (/basePath/app/my-app). Use the `absolute` option to generate an absolute url (http://host:port/basePath/app/my-app)
+Returns the absolute path (or URL) to a given app, including the global base path.
 
-Note that when generating absolute urls, the origin (protocol, host and port) are determined from the browser's location.
+By default, it returns the absolute path of the application (e.g `/basePath/app/my-app`<!-- -->). Use the `absolute` option to generate an absolute url instead (e.g `http://host:port/basePath/app/my-app`<!-- -->)
+
+Note that when generating absolute urls, the origin (protocol, host and port) are determined from the browser's current location.
 
 <b>Signature:</b>
