+++ src/core/public/application/utils/append_app_path.test.ts
@@ -1,0 +1,40 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { appendAppPath } from './append_app_path';
+
+describe('appendAppPath', () => {
+  it('appends the appBasePath with given path', () => {
+    expect(appendAppPath('/app/my-app', '/some-path')).toEqual('/app/my-app/some-path');
+    expect(appendAppPath('/app/my-app/', 'some-path')).toEqual('/app/my-app/some-path');
+    expect(appendAppPath('/app/my-app', 'some-path')).toEqual('/app/my-app/some-path');
+    expect(appendAppPath('/app/my-app', '')).toEqual('/app/my-app');
+  });
+
+  it('preserves the trailing slash only if included in the hash or appPath', () => {
+    expect(appendAppPath('/app/my-app', '/some-path/')).toEqual('/app/my-app/some-path');
+    expect(appendAppPath('/app/my-app', '/some-path#/')).toEqual('/app/my-app/some-path#/');
+    expect(appendAppPath('/app/my-app#/', '')).toEqual('/app/my-app#/');
+    expect(appendAppPath('/app/my-app#', '/')).toEqual('/app/my-app#/');
+    expect(appendAppPath('/app/my-app', '/some-path#/hash/')).toEqual(
+      '/app/my-app/some-path#/hash/'
+    );
+    expect(appendAppPath('/app/my-app', '/some-path#/hash')).toEqual('/app/my-app/some-path#/hash');
+  });
+});
